# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "AMIDDOCS OFFICIAL SWAGGER",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/requests/by-department-id": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "get all description by dep id from query requires secretary access",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "DepartmentRequests",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "department id",
                        "name": "depId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqmodel.RequestViewDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/requests/cancel": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "\"Delete request if status == SEND\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Cancel request by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "requestId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/requests/generate-document": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "returns document from request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "GenerateDocumentFromRequest",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "reqId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqmodel.RequestViewDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/requests/my-requests": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "get user requests by id from jwt token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "UserRequests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqmodel.RequestDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/requests/send": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "\"Send request to generate document\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Send request",
                "parameters": [
                    {
                        "description": "request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/reqmodel.CreateRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/reqmodel.RequestDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/requests/set-done": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    },
                    {
                        "Token": []
                    }
                ],
                "description": "set request status done by req id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "SetRequestStatusDone",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "request id",
                        "name": "reqId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/reqmodel.RequestViewDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/amiderrors.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "amiderrors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
...
        "doctypefields.DocumentType": {
            "type": "string",
            "enum": [
                "STUDY_DOCUMENT_BUDGET",
                "STUDY_DOCUMENT_NO_BUDGET"
            ],
            "x-enum-varnames": [
                "STUDY_DOCUMENT_BUDGET",
                "STUDY_DOCUMENT_NO_BUDGET"
            ]
        },
        "doctypemodel.DocumentTypeDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "refresh_time": {
                    "type": "integer"
                },
                "role": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userfields.Role"
                    }
                },
                "type": {
                    "$ref": "#/definitions/doctypefields.DocumentType"
                }
            }
        },
...
        "reqfields.Status": {
            "type": "string",
            "enum": [
                "SEND",
                "IN_PROGRESS",
                "DONE"
            ],
            "x-enum-varnames": [
                "SEND",
                "IN_PROGRESS",
                "DONE"
            ]
        },
        "reqmodel.CreateRequestDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "departmentId": {
                    "type": "integer"
                },
                "documentType": {
                    "$ref": "#/definitions/doctypefields.DocumentType"
                }
            }
        },
        "reqmodel.RequestDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "departmentId": {
                    "type": "integer"
                },
                "documentType": {
                    "$ref": "#/definitions/doctypemodel.DocumentTypeDTO"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/reqfields.Status"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "reqmodel.RequestViewDTO": {
            "type": "object",
            "properties": {
                "documentCount": {
                    "type": "integer"
                },
                "documentType": {
                    "$ref": "#/definitions/doctypefields.DocumentType"
                },
                "fio": {
                    "$ref": "#/definitions/reqmodel.fio"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/reqfields.Status"
                }
            }
        },
        "reqmodel.fio": {
            "type": "object",
            "properties": {
                "fatherName": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
        "userfields.Role": {
            "type": "string",
            "enum": [
                "STUDENT",
                "ADMIN",
                "SECRETARY"
            ],
            "x-enum-varnames": [
                "STUDENT",
                "ADMIN",
                "SECRETARY"
            ]
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Bearer Token Auth",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "Token": {
            "description": "Client Token auth",
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}